refactor：add Interface encapsulation
增加接口封装，
改变函数的返回值等小修改
